Fixes
Portal bug slow to change gravity bug fixed
Player being pushed when moving along a platform bug fixed for NORMAL
ONLY
New bug created. If one frame you are lower than a platform, and on the
next frame you are on the platform but PLAYER_SPEED, you will be pushed
off. This bug will now be referred to as a feature.